icon-view: Port to GtkFlowBox
This is NOT the final version of the port.

I am glad about the similarities between the ListBox and the
FlowBox API, to the point that I now realize that it would
be more elegant to implement a generic View and leave the
specific bits to the Boxes.IconView and Boxes.ListView
classes only.

Other than that, the Collection class could be a GListModel,
which would apply the sorting and filtering independently
from the views.

All in all, with this patch you should be able to use all the
IconView features as they were before with libgd's MainView.

This is part of an initiative to move away from libgd in cases
where we have a compatible widget in Gtk+.

Stay tuned, in this same bat-repository, same bat-branch.